Squashed 'CadastroServer/' changes from 90f4703..de57d49
de57d49 V2 Servidor Completo
REVERT: 90f4703 Delete Relatório do primeiro procedimento da Missão Prática.pdf
REVERT: afd95ea Add files via upload

git-subtree-dir: CadastroServer
git-subtree-split: de57d491676e91067a79d8050a5d7a060b65fa8d

+++ src/model/Movimento.java
@@ -8,6 +8,7 @@ import java.io.Serializable;
 import javax.persistence.Column;
 import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
+import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.NamedQueries;
@@ -39,7 +40,7 @@ public class Movimento implements Serializable {
     @Column(name = "tipo")
     private String tipo;
     @Column(name = "valorUnitario")
-    private Long valorUnitario;
+    private Integer valorUnitario;
     @JoinColumn(name = "Pessoa_idPessoa", referencedColumnName = "idPessoa", insertable = false, updatable = false)
     @ManyToOne(optional = false)
     private Pessoa pessoa;
@@ -52,7 +53,7 @@ public class Movimento implements Serializable {
 
     public Movimento() {
     }
-
+   
     public Movimento(MovimentoPK movimentoPK) {
         this.movimentoPK = movimentoPK;
     }
@@ -85,11 +86,11 @@ public class Movimento implements Serializable {
         this.tipo = tipo;
     }
 
-    public Long getValorUnitario() {
+    public Integer getValorUnitario() {
         return valorUnitario;
     }
 
-    public void setValorUnitario(Long valorUnitario) {
+    public void setValorUnitario(Integer valorUnitario) {
         this.valorUnitario = valorUnitario;
     }
